+++ protostar/protostar_toml/protostar_contracts_section.py
@@ -2,15 +2,15 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
-from protostar.protostar_toml.core import (
-    InvalidProtostarTOMLException,
-    ProtostarTOMLReader,
-    ProtostarTOMLSection,
-)
+from protostar.protostar_toml.protostar_toml_exceptions import \
+    InvalidProtostarTOMLException
+from protostar.protostar_toml.protostar_toml_reader import ProtostarTOMLReader
+from protostar.protostar_toml.protostar_toml_section import \
+    ProtostarTOMLSection
 
 
 @dataclass
-class ProtostarContracts(ProtostarTOMLSection):
+class ProtostarContractsSection(ProtostarTOMLSection):
     contract_dict: Dict[str, List[Path]]
 
     @staticmethod
@@ -20,14 +20,14 @@ class ProtostarContracts(ProtostarTOMLSection):
     @classmethod
     def from_protostar_toml(
         cls, protostar_toml: ProtostarTOMLReader
-    ) -> "ProtostarContracts":
+    ) -> "ProtostarContractsSection":
         section_dict = protostar_toml.get_section(cls.get_section_name())
         if section_dict is None:
             raise InvalidProtostarTOMLException(cls.get_section_name())
         return cls.from_dict(section_dict)
 
     @classmethod
-    def from_dict(cls, raw_dict: Optional[Dict[str, Any]]) -> "ProtostarContracts":
+    def from_dict(cls, raw_dict: Optional[Dict[str, Any]]) -> "ProtostarContractsSection":
         if not raw_dict:
             return cls(contract_dict={})
