+++ protostar/protostar_toml/protostar_project_section.py
@@ -2,15 +2,15 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Dict
 
-from protostar.protostar_toml.core import (
-    InvalidProtostarTOMLException,
-    ProtostarTOMLReader,
-    ProtostarTOMLSection,
-)
+from protostar.protostar_toml.protostar_toml_exceptions import \
+    InvalidProtostarTOMLException
+from protostar.protostar_toml.protostar_toml_reader import ProtostarTOMLReader
+from protostar.protostar_toml.protostar_toml_section import \
+    ProtostarTOMLSection
 
 
 @dataclass
-class ProtostarProject(ProtostarTOMLSection):
+class ProtostarProjectSection(ProtostarTOMLSection):
     libs_path: Path
 
     @staticmethod
@@ -20,14 +20,14 @@ class ProtostarProject(ProtostarTOMLSection):
     @classmethod
     def from_protostar_toml(
         cls, protostar_toml: ProtostarTOMLReader
-    ) -> "ProtostarProject":
+    ) -> "ProtostarProjectSection":
         section_dict = protostar_toml.get_section(cls.get_section_name())
         if section_dict is None:
             raise InvalidProtostarTOMLException(cls.get_section_name())
         return cls.from_dict(section_dict)
 
     @classmethod
-    def from_dict(cls, raw_dict: Dict[str, Any]) -> "ProtostarProject":
+    def from_dict(cls, raw_dict: Dict[str, Any]) -> "ProtostarProjectSection":
         return cls(libs_path=Path(raw_dict["libs_path"]))
 
     def to_dict(self) -> "ProtostarTOMLSection.Dict":
