+++ protostar/protostar_toml/protostar_toml_reader.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Optional
 import flatdict
 import tomli
 
-from protostar.protostar_toml.core.protostar_toml_exceptions import (
+from protostar.protostar_toml.protostar_toml_exceptions import (
     NoProtostarProjectFoundException,
 )
 from protostar.utils.protostar_directory import VersionManager
